requirements.txt
git+https://github.com/ophexor/yt-dl@master
pyrogram
TgCrypto
py-tgcalls==0.8.1rc1
python-dotenv
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg-python
ffmpeg
Pillow
ujson
